// src/core/config/config.d.ts
export declare const config: {
    api: {
        port: string | undefined;
        host: string | undefined;
    };
    postgresql: {
        user: string | undefined;
        password: string | undefined;
        host: string | undefined;
        port: string | undefined;
        database: string | undefined;
    };
};
//# sourceMappingURL=config.d.ts.map